requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.3
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.0
dill==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.7
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.0.1
pandas==2.2.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
pylint==3.2.5
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
